\name{mfpp}
\alias{mfpp}
\docType{package}

\title{Matrix-Based Flexible Project Planning}

\description{
Matrix-Based Flexible Project Planning. This package model, plan and schedule flexible, such as agile, extreme and hybrid project plans. The package contains project planning, scheduling and risk assessment functions.
}

\seealso{\code{\link{generatepdm}}, \code{\link{get.structures}}, \code{\link{is.flexible}}, \code{\link{percent}},\code{\link{phase1}},\code{\link{phase2}},\code{\link{phase3}},\code{\link{plot.PDM_matrix}},\code{\link{summary.PDM_matrix}}.}

\author{Zsolt T. Kosztyan*, Aamir Saghir

*e-mail: kzst@gtk.uni-pannon.hu}

\references{
Kosztyán, Z. T. (2022). MFPP: Matrix-based flexible project planning. SoftwareX, 17, 100973.
}

\keyword{package}
